Fix socket.io change on room's socket list
This change isn't even listed in the changelog, but fatal:
io.sockets.adapter.rooms[<room name>] seems to be an Object of type Room now, having a the list of its sockets in an property "sockets".
Before version 1.4.0, the map of sockets was directly accessable by the call (there was no Room object)...

Moved logic for iterating the room's sockets to a new method _iterateEachJoinedSocket() though. So future changes to the API (which are very likely) can be modified on one place in the code...