Append IRC tag to related users on custom channels
On the global channel, the IRC users are recognizable by the missing of an assigned game.
But on custom channels this is not true, if using the mono-bot.
So now all IRC users on custom channels are marked as IRC - Except for one case:
A game-bot (not a mono-bot) is only marked as IRC, if he doesn't send a normal message - e.g. join, mode-change, etc.