Bugfix: Logger always enabled debug log though broken lazy loading
The logger is imported by dependencies of the datasource loader.
The loader is imported before updating config
Hence, the logger is imported before updating config and the debug is enabled by default...

Fix: Do not depend on lazy loading for the logger. Just check enable flag on every debug log.